lab_4/requirements.txt
cassandra-driver
fastapi
pandas
uvicorn